Fix: Do not double-connect
When AlarmDaemon::ServerSocket->new() is called, a socket is
created and connected.

Scripts are calling AlarmDaemon::ServerSocket->connect() when in
fact a socket already exists and is connected, causing the socket
to be replaced with a new connected socket.

Make it so a call to connect() with an already connected socket is
ignored.