Remove the pending_data stuff
This wasn't so great. Pending Data could be only MINPIC, XINPIC and
CONTACTID lines and it *might* be reasonable to buffer them in the
case of no clients, but they'll never be buffered when there's more
than one client.

Also when pending data is sent, it doesn't come with timestamps.

To make some reliable way for a client to ensure that it doesn't miss
any pending data would require numbered streams or event times, and
an extension to the protocol.