Make automatic reconnection work
If connect fails, call the defined onConnectFail subroutine. This
works except for initial connection within new(). To compensate
for the bad API, calling onConnectFail to set the sub while not
connected will cause the sub to be called (and thence, retry).

If connected and a disconnection occurs, call the defined onDisconnect
subroutine.

Exponential backoff works nicely.

All this means the daemon can now start up if its server is not
presently available and it will reconnect automatically after a
failure.

Queueing of send commands while awaiting connection is not yet sorted.