Remove CRLF from sent commands
When sending a command upstream, don't suffix it with CRLF. Turns out
the LS-30 behaves slightly better, by not sending a !& null response after
the response.

Correspondingly, commands from downstream are complete when a '&' is
received. Buffered data which cannot form a command must be deleted
from the buffer. This includes CR LF.