Improve simulator: eventlog, date/time
Add a dummy return string to event log queries

Store updated settings in the state file.

If date/time set is requested, calculate an offset between the
current time and the set time, and store it in the state file.
Correspondingly, when date/time is queried, apply the offset
to the current time before returning a date/time string.

Add support for the unknown 'l5' command.